added store and reducer
