+++ client/package.json
@@ -9,7 +9,9 @@
     "react": "^16.10.1",
     "react-beautiful-dnd": "^11.0.5",
     "react-dom": "^16.10.1",
+    "react-redux": "^7.1.1",
     "react-scripts": "3.1.2",
+    "reselect": "^4.0.0",
     "styled-components": "^4.4.0"
   },
   "scripts": {
