+++ client/src/actions/consts.js
@@ -1,0 +1,1 @@
+export const CHANGE_ROUTE = 'CHANGE_ROUTE'
