+++ client/src/actions/index.js
@@ -1,0 +1,5 @@
+import {CHANGE_ROUTES} from './consts'
+
+export const addArticle = (payload)=>{
+    return { type:CHANGE_ROUTES, payload}
+}
