+++ client/src/index.js
@@ -3,8 +3,17 @@ import ReactDOM from 'react-dom';
 import './index.css';
 import * as serviceWorker from './serviceWorker';
 import Dashboard from './components/Dashboard';
+import {Provider} from 'react-redux'
+import store from './store/index'
 
-ReactDOM.render(<Dashboard />, document.getElementById('root'));
+
+
+const rootElement = document.getElementById('root')
+ReactDOM.render(
+    <Provider store={store}>
+        <Dashboard />
+    </Provider>
+, rootElement);
 
 // If you want your app to work offline and load faster, you can change
 // unregister() to register() below. Note this comes with some pitfalls.
