+++ client/src/store/index.js
@@ -1,0 +1,6 @@
+import {createStore} from 'redux'
+import routesReducer from '../reducers/routes'
+
+const store = createStore(routesReducer)
+
+export default store
